packages/bub-discord/pyproject.toml
[project]
name = "bub-discord"
version = "0.1.0"
description = "Discord channel for Bub"
readme = "README.md"
authors = [
    { name = "Frost Ming", email = "me@frostming.com" }
]
requires-python = ">=3.12"
dependencies = [
    "bub>=0.3.0",
    "discord-py>=2.7.1",
]

[project.entry-points.bub]
discord = "bub_discord.plugin"

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"
